backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
